Refactor and move the lemma function which copies the conllu files to prepare_tokenizer_treebank, since the POS copy will be almost exactly the same
